# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VOTES-API Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/dashboard/candidates/most-voted": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Obtener los candidatos más votados",
                "description": "Devuelve una lista de los candidatos más votados con su información y el conteo de votos",
                "operationId": "b9ac70ffab8cdc9b95ba1227fe76703a",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token de autorización Bearer",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer {token}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de candidatos más votados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "John"
                                            },
                                            "lastname": {
                                                "type": "string",
                                                "example": "Doe"
                                            },
                                            "votes": {
                                                "type": "integer",
                                                "example": 150
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}